Fixed three bugs that manifested under load:
1. We were not freeing I/O buffers when we were done with them!

2. And when I added a free() in shutdownConnection(), the server
started crashing with SEGVs. That's because shutdownConnection()
was being called from the wrong place: the connection has to stay
alive until the last uring write operation completes, so the
shutdown has to happen during write CQE completion processing.

3. Finally, some connections just weren't getting serviced.
If MAX_CONNECTIONS==128, and I tried 100 simultaneous fetches
of the same file, most of them would complete but some would
never even start transferring. That was because I had set the
socket listen backlog to 8. Raising it to 100 seems to have
fixed the problem.